provide JPQL query for aggregate statistic

+++ src/main/java/pl/pawel/cqrs/persistence/aggregatestatistic/PersonAggregateStatistics.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class PersonAggregateStatistics {
 
     private String name;
-    private int avgSalary;
+    private double avgSalary;
 }
